Cleanup and exit after recieving SIGINT
If restic receives SIGINT, it performs cleanup operations and exits with
code 130.Bash doesn't recognise this as having been killed by a signal,
so keeps running (ignoring its own SIGINT). This commit add exit code
checks to fix that.

It also ensures that snapshots are properly deleted if the script exits
in response to any trappable signal.